Flip the CCD layout left-to-right to provide an on-sky view
This is consistent with the camera offsets, and means that the WCS is
no longer backwards.  It requires that we flip the pixels too, and this
is done in the mapper.

N.b. we don't (yet) flip the bias/data definitions, but as they are
in fact left-right symmetric this is OK.