SConscript: generate defects automatically
Since we want to use the defects, we need to generate them automatically.
I wasn't able to work out how to do them conditionally (one input file
makes many output files, and I don't know which ones), so it's always run
(it's not too expensive).
This requires some mucking around with the environment so the python scripts
we're calling can import things like numpy, pyfits.  Messy, but it works.