isr: disable multiplication by gain
Gain differences from CCD to CCD should be taken out by the flat-field,
not by multipying by the gains (because the flats measure the relative
differences better than gains can).

I'm sure I fixed this a couple years ago, but it seems to have fallen out.
Thanks to Yasuda-san for finding this problem.