Adds a utility for building camera registries
The old camera objects were built from policy files.  This
is no longer required, but since the representation was available
we continue to use the policy file as the starting point for
building camera registries (the persisted files used by the mapper
to build a camera).  Also, the transform registry is different
between hsc and suprimecam.  This makes the repo generation
respect this.