Adds comment about SUBARU_DATA
I don't know where to find the package the mapper expects
as default called subaru_data.  I added $OBS_SUBARU_DIR
as the default instead of $SUBARU_DATA_DIR.  This will need
to be addressed.  This also removes std_camera since the
use of std_camera was re-defined in the current master
CameraMapper.