Remove swig in lsst/obs/hsc
I had moved the swig wrappings to lsst/obs/subaru because I thought
it might be nice to have all the swigged code in one place.  In
doing so, I gutted a file but left it haning around.  This commit
removes the no-op .i file and the SConstruct file that built it.