Add strayFlux option to the deblender to shrink parents to the union of their children.
The deblender is not doing well at assigning stray flux to its children.  This
option ignores pixels not assigned to any children and trims the parent footprint to the
union of the children.  This can result in parents having non-contigous footprints.

Conflicts:
	python/lsst/meas/deblender/baseline.py

This is a cherry-pick from HSC-1237's commit:
https://github.com/HyperSuprime-Cam/meas_deblender/commit/d509aee5a12496025293666555dab15aae32b4a4