use Image instead of MaskedImage for templates
This is intended to save memory, and though it's not as effective as
originally hoped, yet provides a useful savings.  The variance plane of
the templates was completely unused, and the mask plane was only used for
debugging that no-one ever looked at (and has since been removed to save
the additional mask planes).

Conflicts:
	python/lsst/meas/deblender/baseline.py
	src/Baseline.cc

This is a cherry-pick from HSC-1265's commit:
https://github.com/HyperSuprime-Cam/meas_deblender/commit/bdd0bfb35ad2c70588869e736f6f02bd6dfdafa8