HscMapper: add dataset to indicate reprocessing is underway
This is part of a workaround for large deblender footprints causing
large memory use and/or very slow processing.  We refuse to deblend
those footprints when running on a cluster and return to reprocess
on a machine with larger memory or more time if we consider those
footprints important to recover.  This file indicates that that
reprocessing is underway, so that if the reprocessing fails we can
identify patches that still require reprocessing.