clip template images in addition to footprints
When clipFootprintToNonzero (True by default, and not modified), the
footprints were being clipped where the image was non-zero, but the
images (which carry most of the memory) were not, so there was no
memory savings.  It's simple to use the bounding box on the clipped
footprint to make a new, smaller template image, to realise a decent
memory savings.