update parameters limiting size of footprints to be deblended
After reducing the memory footprint, we can tolerate a maximum footprint
size of 10^6 pixels without violating the memory limit of about 4 GB.

Removed the limit on the linear size of the footprint and the axis ratio.
These were originally intended to remove large satellite streaks, but we
are removing those in other ways (clipping in coadd, maybe some dedicated
streak detection), and the real operational constraint is on the total
number of pixels rather than the shape.