Optimize getSignificantEdgePixels and hasSignificantFluxAtEdge
In order to find edge pixels, was calling Footprint::contains twice for
every pixel, which does a linear search and is therefore slow for large
Footprints.  Optimized this and moved it to afw as
Footprint::findEdgePixels.  Then we just have to find which pixels on the
edge are significant.

Conflicts:
	src/Baseline.cc

This is a cherry-pick from HSC-1274's commit:
https://github.com/HyperSuprime-Cam/meas_deblender/commit/4d93e97201cfb639bf4a4f566dd5648ebbabc894