HscMapper: update to add flags back in for butler interfacing
The 'flags' argument must be passed to the butler (it is used to
change how the reading/writing is done).  However, it must be removed
from the mapper (because it doesn't correspond to a registry column).
This updates the map function such that the 'flags' argument is added
back to the dataId for future butler interfacing.