Update config override files new default astrometry task
The default astrometry task used by CalibrateTask is now
AstrometryTask instead of ANetAstrometryTask. I changed
two config override files to compensate for this:

* processStack.py does not set astrometry.solver.calculateSip.
This was required, and the fix is to delete the line
because AstrometryTask will not try to fit SIP terms
if it does not fit a WCS)

* processCcd.py does not retarget to use AstrometryTask.
This was optional, since it does not change behavior,
but it simplifies the file.