Update processCoadd config to bypass cmodel
The cmodel fitting code is undergoing a significant overhaul.  Given
the large runtime and occasional failures, we here turn off cmodel
while we await these latest fixes (large blends, footprint merging,
etc.) are in.

This is an update to the following commit where cmodel was turned
off in the other config override files:
https://github.com/lsst/obs_subaru/commit/2874722faf7c2a601f5f286c868ecd810127befb