Add a mapper between the HSC/LSST src schemas
This provides a mapping between schema names used on what is currently the
"HSC stack" and the main LSST stack.  The former will eventually become
obsolete, but for now this mapper is required for making comparisons between
outputs produced by the different stacks.

Note that this list is not exhaustive.