Fix rotation for isr correction
Approximately half of the HSC CCDs are rotated 180 deg with respect
to the others. Two others have 90 deg rotations and another two have
270 deg rotations (see HSC CCD layout).  The raw images for the
rotated CCDs thus need to be rotated to match the rotation of their
associated calibration frames prior to applying the corrections.
This is accomplished by rotating the exposure using the rotated
context manager function in obs_subaru's isr.py and the nQuarter
specification in the policy file for each CCD. Currently, rotated
uses afw's rotateImageBy90 (which apparently rotates in a counter-
clockwise direction) to rotated the exposure by 4 - nQuarter turns.
This turns out to be the wrong rotation for the odd nQuarter CCDs.
This fixed the rotate function to perform the correct rotation prior
to the given isr.