Fix effective coordinates for defects
The defects in obs_subaru are defined assuming that pixel (0, 0) is
the lower-left corner.  However, the LSST stack does not use this
interpretation, preferring to maintain the coordinate system tied
to the electronics.  As such, the defect positions are being
misinterpreted for the rotated CCDs in HSC so we need to rotate the
defects bbox in order to apply the badAmp assignment to the correct
amplifier.