Update setting of badAmp for dead amplifiers
The defects in obs_subaru are defined assuming that pixel (0, 0) is the
lower-left corner.  This is not true for the rotated HSC ccds, so we
need to rotate the defects bbox in order to apply the badAmp assignment
to the correct amplifier.  This rotation needs to be done in
trimmed/assembled image dimensions as that is the frame for which the
defects are defined.

Also, removed the overridden isr saturationDetection() function as
it is no longer necessary as of 2f32ea9bfc.