specify defects in electronic frame
Defects have been specified in the coordinate frame of the physical
camera (due to the former cameraGeom rotating all images to be in the
phyiscal camera frame), which required rotating them before application,
and would confuse anyone who attempted to add new defects. Fixed this
situation by applying the rotation to the defects (using the code from
isr.py that used to do the rotation) to produce new defects lists in the
(unrotated) electronics frame (as the images come off the telescope, but
including the flip in reading the raw data done in HscMapper.std_raw)
that is now standard. Removed the rotation in isr.py since that is no
longer needed.