Avoid config error with empty astrometry_net_data.
setConfigFromEups() throws an error if the setup version of astrometry_net_data
is not found in the menu. This occurs even if the user disables photometric and
astrometric calibration. Adding a glob pattern for the empty astrometry_net_data
version allows the config overrides to run without a real reference catalog.