config: fix base_ClassificationExtendedness value
This parameter should be set to 0.95 when GaussianFlux is being used
as ModelFlux.  If the package for CModel is set up (i.e CModel is
being used as ModelFlux), the change to 0.985 appropriate for CModel
is done in the configuration file that enables CModel.  Thus the
config setting in processCcd.py should be 0.95.