activate measurements used in s16a production afterburner
The HSC s16a production's afterburner included two measurements features
which have been ported to LSST: measurement on the undeblended image (to
allow potentially useful measurements where the deblender fails, and to
compare with software that doesn't use a deblender), and measurement of
convolved fluxes (to provide fluxes in multiple bands with a common
seeing, e.g., for photometric redshifts). This patch activates these
features for HSC and Suprime-Cam.