Move variance update after bias correction
With the removal of the requirement that assembly take images with a variance plane
this allows for amp assembly to be done in a more correct order.  This also means
that at the point that the variance plane is updated the assembled ccd has been trimmed
of prescan/overscan regions.