Add explicit multiplier to CModel likelihood.
This *approximately* reverts the fix made to meas_modelfit on
DM-9795 by applying an explicit constant scaling (instead of an
accidental, variance-dependent one) to the likelihood.

Hopefully we can revert this change entirely (or at least make
the scaling closer to one) once we check that a less informative
prior doesn't cause problems.