HscMapper: cache the camera object
The lion's share of the time in instantiating a butler goes to
building the camera object, because that involves parsing a Config
with many lines, and a stat call for each line to get the traceback.
We cache the camera once it's built so that subsequent butler
instantiations (e.g., for multiprocessing or pipe_drivers) don't
suffer the same overhead.