config: disable old behaviour of CModel
DM-9795, fixed a bug in the relative weight of likelihood vs prior,
but configuration parameters were added here to retain (approximately)
the old behaviour. We now believe that the new behavior is far superior
to the old and doesn't appear to introduce any ill effects, so the
configuration changes have been removed to allow default use of the
fixed CModel.