config/hsc/processCcd.py: update maxOffsetPix
The expansion of the bbox search radius for reference objects to
bbox + maxOffsetPix (added in DM-9750) caused a significant regression
in the astrometric matching.  It seems this is due to an issue with how
matchOptimisticB is handling large numbers of reference stars.  The
default in obs_subaru of maxOffsetPix = 750 resulted in a significant
number of CCDs failing in processCcd (identified in DM-11020).  Here we
set that default to maxOffsetPix = 250 which effectively results in the
same number of CCDs in the RC dataset passing single frame processing
and equivalent quality based on a comparison of QA plots for the pre-
regression and post-fix runs.