HscMapper: always set calibRoot
If calibRoot is not set, we can't get the registry, and if we can't get
the registry then we can't get calibs. This is now necessary since we're
not allowed to fall over to a PosixRegistry if the calibRoot doesn't exist
(DM-10443). So, try our hardest to find the calibRoot by searching the
parents that are provided to us.