Revert "config: allow EDGE pixels to be coadded"
This reverts commit 5faadffd2f408b019280ea1c353f0ec787b092b9.

We've identified a reason to ignore EDGE pixels in the coadd:
the charge from saturated sources can pool at the top/bottom
of the detector. This charge is at a level (~20k) significantly
below saturation, so it's difficult to identify cleanly. Tossing
all pixels around the edge of the image is the easiest way to
ensure these don't leak in.