config: disable apCorr registration of undeblended convolved measurement
This suppresses warnings of the type:

    Cannot aperture correct undeblended_ext_convolved_ConvolvedFlux_1_6_0 because could not find undeblended_ext_convolved_ConvolvedFlux_1_6_0_flux or undeblended_ext_convolved_ConvolvedFlux_1_6_0_fluxSigma in apCorrMap

Because the columns used by the convolved measurement are created at
plugin instantiation (because the names are defined by the config),
registration for the aperture correction is done at plugin instantiation
time too. But we don't want them registered for the undeblended
measurement because we'll get the aperture correction from the deblended
proxy.