Disable subimage access to raws.
Gen2 seems to have have always done this incorrectly given that
the flipping that happens in std_raw should have been considered
when interpreting the bounding boxes.  In Gen3 we'll instead refuse
to do it (unless we decide we need it later).

This does slow down a unit test (because we need to read a full file
in it).