Fix missing INST-PA message for y-band data.
The rewrite of ISR to be more general also attempted to remove as many
extra butler calls as possible.  However, as ci_hsc does not test
y-band data, this caused a hidden bug with the y-straylight
correction.  The required keywords should not be pulled from the
exposure header, but rather from the VisitInfo associated with the
exposure.