Update defineVisits padding override for Gen3
The original estimate of 4000 was based old "raw" WCS definitions that
did not take the full boresight + rotation + cameraGeom (distortion)
information into account and old versions of the matchers and fitters
used in astrometric calibration.  This number should be a small as
possible while still ensuring the padding covers the full calibrated
footpring for all visits.  A reassessment based on the RC2 dataset sees
a maximum (jointcal - raw) pixel offset (where only the CCD corners were
considered) of ~191 pixels (in tract coordinates with 0.168 arcsec/pixel).
This would indicate we might be able to go as low as 200 pixel padding.
This analysis is continuing on a larger dataset (the full HSC PDR2) and
initial indications point to some slightly higher offsets: the largest
seen thus far is ~220 pixels.  Thus, to be somewhat conservative, we
adopt a padding of 250 pixels (to be reconsidered in the context of the
full PDR2 analysis being conducted on DM-28011.