Update Gen2 calibration policy formats
The Gen2 obs_subaru policy/HscMapper.yaml defines bias and dark calibrations
to be of persistence type ImageF which, by definition, does not have mask or
variance planes.  On read, these are passed through a standardization step
that converts ImageF to an ExposureF by adding empty mask and variance planes.
However, as revealed during a test processing the same data in Gen3, it was
noted that these planes do indeed exist and are being used and propagated
throughout the Gen3 processing.  While the ommission in Gen2 processing is
likely scientifically insignificant, we should still be using them for a
rigorous accounting for and propagation of know sources of measurement errors.

The more immediate desire for this change here is that having such (albeit
minor) differences puts a dent in our efforts towards acheiving Gen2/Gen3
parity (toward the imminent retirement of the former) through Single Frame
Processing. The fix here is to update the bias and dark formats in the policy
file from ImageF to MaskedImageF.