Loosen initial astrometry match distance threshold
The minMatchDistanceArcSec config is the maximum match distance (in
arcsec) below which further iteration is pointless.  Since we can only
do as well as the camera model in SFM using the affine WCS fit, the
default value is way too strict for our currently defined camera model.
This can be tightened up (or removed) if an updated camera model is
implemented.  Further refinement to the WCS is left for the subsequent
global fit (namely gbdes).